+++ src/UnitTestEx.NUnit/Internal/GenericTesterT.cs
@@ -1,7 +1,6 @@
 ﻿// Copyright (c) Avanade. Licensed under the MIT License. See https://github.com/Avanade/UnitTestEx
 
 using UnitTestEx.Generic;
-using NFI = NUnit.Framework.Internal;
 
 namespace UnitTestEx.NUnit.Internal
 {
@@ -13,6 +12,6 @@ namespace UnitTestEx.NUnit.Internal
         /// <summary>
         /// Initializes a new instance of the <see cref="GenericTester"/> class.
         /// </summary>
-        internal GenericTester() : base(new NUnitTestImplementor(NFI.TestExecutionContext.CurrentContext)) { }
+        internal GenericTester() : base(new NUnitTestImplementor()) { }
     }
 }