+++ src/UnitTestEx/Abstractions/TestSharedState.cs
@@ -1,7 +1,9 @@
 ﻿// Copyright (c) Avanade. Licensed under the MIT License. See https://github.com/Avanade/UnitTestEx
 
 using CoreEx.Events;
+using Microsoft.Extensions.Logging;
 using System.Collections.Concurrent;
+using System.Collections.Generic;
 using UnitTestEx.Expectations;
 
 namespace UnitTestEx.Abstractions
@@ -11,6 +13,9 @@ namespace UnitTestEx.Abstractions
     /// </summary>
     public sealed class TestSharedState
     {
+        private readonly object _lock = new();
+        private readonly List<string?> _logOutput = new();
+
         /// <summary>
         /// Initializes a new instance of the <see cref="TestSharedState"/> class.
         /// </summary>
@@ -22,8 +27,34 @@ namespace UnitTestEx.Abstractions
         public ConcurrentDictionary<string, ConcurrentQueue<EventData>> EventStorage { get; } = new();
 
         /// <summary>
-        /// Resets the shared state.
+        /// Adds the <see cref="ILogger"/> log message.
         /// </summary>
-        public void Reset() => EventStorage.Clear();
+        /// <param name="message">The log message.</param>
+        public void AddLoggerMessage(string? message)
+        {
+            lock (_lock)
+            {
+                _logOutput.Add(message);
+            }
+        }
+
+        /// <summary>
+        /// Gets the <see cref="ILogger"/> messages.
+        /// </summary>
+        /// <remarks>Also clears the messages.</remarks>
+        public IEnumerable<string?> GetLoggerMessages()
+        {
+            lock (_lock)
+            {
+                var logs = _logOutput.ToArray();
+                _logOutput.Clear();
+                return logs;
+            }
+        }
+
+        /// <summary>
+        /// Resets the <see cref="EventStorage"/> shared state.
+        /// </summary>
+        public void ResetEventStorage() => EventStorage.Clear();
     }
 }