+++ src/UnitTestEx/Generic/GenericTesterBase.cs
@@ -3,6 +3,7 @@
 using CoreEx;
 using Microsoft.Extensions.DependencyInjection;
 using System;
+using System.Linq;
 using System.Threading.Tasks;
 using UnitTestEx.Abstractions;
 using UnitTestEx.Assertors;
@@ -72,8 +73,6 @@ namespace UnitTestEx.Generic
 
             Implementor.WriteLine("VALIDATE >");
             Implementor.WriteLine("Validator: <function>");
-            Implementor.WriteLine("");
-            Implementor.WriteLine("LOGGING >");
 
             var sw = System.Diagnostics.Stopwatch.StartNew();
 
@@ -91,6 +90,19 @@ namespace UnitTestEx.Generic
             }
 
             sw.Stop();
+
+            Implementor.WriteLine("");
+            Implementor.WriteLine("LOGGING >");
+            var messages = SharedState.GetLoggerMessages();
+            if (messages.Any())
+            {
+                foreach (var msg in messages)
+                {
+                    Implementor.WriteLine(msg);
+                }
+            }
+            else
+                Implementor.WriteLine("None.");
 
             Implementor.WriteLine("");
             Implementor.WriteLine("RESULT >");